sandbox: fix compatibility issues
1. Podsandbox controller do not store the sandbox itself, the sbserver
store the sandbox in the last step of RunPodSandbox, so we can not get
the status of sandbox before store it into the cri sandbox store. we
change the sandbox API to add task_address to controller Start response
so that we don't need to call Status of sandbox when in RunPodSandbox.

2. If we restart containerd with new version with sandbox controller
enabled, the old sandbox which is stored as a container with a special
metadata, should init a Sandbox field in the Sandbox stored in CRI
sandbox store, otherwise the call to the SandboxInstance will come aross
a nil pointer error.